Change all doubles to floats
This should help to reduce code size and improve performance.

+++ src/imu.c
@@ -20,7 +20,7 @@ uint16_t accel_raw[3]; //possibly can be deined in the imu_read function
 uint16_t gyro_raw[3];
 
 // Variables 
-double pitch = 0, roll = 0, pitch_accel = 0, roll_accel = 0;
+float pitch = 0, roll = 0, pitch_accel = 0, roll_accel = 0;
 
 uint8_t write_IMU_byte(uint8_t IMU_address, uint8_t address, uint8_t data_8bit){
 	uint8_t IMU_address_write = (IMU_address << 1);
@@ -277,11 +277,11 @@ void calibrate_IMU(void){
 
 	//set the initial angle according to the accelerometer readings
 	read_raw_accel(&accel_raw[0]);
-	double accel_x = (int)accel_raw[0]/(double)ACCEL_SENSITIVITY;
-	double accel_y = (int)accel_raw[1]/(double)ACCEL_SENSITIVITY;
-	double accel_z = (int)accel_raw[2]/(double)ACCEL_SENSITIVITY;
-	roll = -asin(accel_x/sqrt(accel_x*accel_x+accel_y*accel_y+accel_z*accel_z))*180/3.1415;
-	pitch = asin(accel_y/sqrt(accel_x*accel_x+accel_y*accel_y+accel_z*accel_z))*180/3.1415;	
+	float accel_x = (int)accel_raw[0]/(float)ACCEL_SENSITIVITY;
+	float accel_y = (int)accel_raw[1]/(float)ACCEL_SENSITIVITY;
+	float accel_z = (int)accel_raw[2]/(float)ACCEL_SENSITIVITY;
+	roll = -asinf(accel_x/sqrtf(accel_x*accel_x+accel_y*accel_y+accel_z*accel_z))*180/3.1415;
+	pitch = asinf(accel_y/sqrtf(accel_x*accel_x+accel_y*accel_y+accel_z*accel_z))*180/3.1415;	
 }
 
 void reset_IMU(){
@@ -365,14 +365,14 @@ void read_raw_accel(uint16_t *accel){
 // This function has to be executed at 250Hz frequeny (every 4ms)
 void imu_read(measured_state_t *destination){
 	read_raw_gyro(&gyro_raw[0]);
-	double gyro_x = (int)gyro_raw[0]/(double)GYRO_SENSITIVITY;//-gyroBias[0];
-	double gyro_y = (int)gyro_raw[1]/(double)GYRO_SENSITIVITY;//-gyroBias[1];
-	double gyro_z = (int)gyro_raw[2]/(double)GYRO_SENSITIVITY;//-gyroBias[2];
+	float gyro_x = (int)gyro_raw[0]/(float)GYRO_SENSITIVITY;//-gyroBias[0];
+	float gyro_y = (int)gyro_raw[1]/(float)GYRO_SENSITIVITY;//-gyroBias[1];
+	float gyro_z = (int)gyro_raw[2]/(float)GYRO_SENSITIVITY;//-gyroBias[2];
 
 	read_raw_accel(&accel_raw[0]);
-	double accel_x = (int)accel_raw[0]/(double)ACCEL_SENSITIVITY;//-accelBias[0];
-	double accel_y = (int)accel_raw[1]/(double)ACCEL_SENSITIVITY;//-accelBias[1];
-	double accel_z = (int)accel_raw[2]/(double)ACCEL_SENSITIVITY;//-accelBias[2];
+	float accel_x = (int)accel_raw[0]/(float)ACCEL_SENSITIVITY;//-accelBias[0];
+	float accel_y = (int)accel_raw[1]/(float)ACCEL_SENSITIVITY;//-accelBias[1];
+	float accel_z = (int)accel_raw[2]/(float)ACCEL_SENSITIVITY;//-accelBias[2];
 	
 	/*
 	// Integration of the roll and pitch 
@@ -394,7 +394,7 @@ void imu_read(measured_state_t *destination){
 	// or I am doing something wrong
 
 	MadgwickAHRSupdateIMU(gyro_x, gyro_y, gyro_z, accel_x, accel_y, accel_z);
-	roll= (acosf(q0/sqrt(q0*q0+q2*q2))*2.0f)*180/3.14;
+	roll= (acosf(q0/sqrtf(q0*q0+q2*q2))*2.0f)*180/3.14;
 	pitch= atan2f(2*(q0*q1+q2*q3), 1-2*(q1*q1+q2*q2))*180/3.14;
 	destination->roll = roll;
 	destination->pitch = pitch;
